% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tters-package.R
\docType{package}
\name{tters-package}
\alias{tters}
\alias{tters-package}
\title{tters: Sequential Target Trial Emulation Data Expansion}
\description{
A thin R binding over the verified \code{tte-expand} Rust + Polars engine.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/oldschoolcool2/rust-tte}
  \item Report bugs at \url{https://github.com/oldschoolcool2/rust-tte/issues}
}

}
\author{
\strong{Maintainer}: Michael Batech \email{oldschoolcool@gmail.com}

Authors:
\itemize{
  \item Michael Batech \email{oldschoolcool@gmail.com}
}

}
\keyword{internal}
